Decode MacCapa bit field and move it out
